rename package to http.authentication
